requirements.txt
torch
pillow
numpy
albumentations
tqdm
torchvision
matplotlib